requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.10.6
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
enum34==1.1.6
funcsigs==1.0.2
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
Markdown==2.6.8
mock==2.0.0
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
python-decouple==3.0
python-http-client==2.2.1
scandir==1.5
sendgrid==3.6.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
